more cleaning/refactoring/documentation
